wip: added headers

+++ src/main/java/com/chimpler/simtick/Version.java
@@ -2,5 +2,5 @@ package com.chimpler.simtick;
 
 public class Version {
     // version will change when reader/writer definitions change
-    public static final int DataVersion = 1;
+    public static final short DataVersion = 1;
 }
